requirements.txt
fastapi
uvicorn[standard]
python-dotenv
docling
langchain-text-splitters
sentence-transformers
qdrant-client
openai
python-multipart
langchain
langchain-core
langchain-ollama
langchain-openai
langgraph
rank-bm25